requirements.txt
Package              Version
-------------------- -----------
﻿requests==2.28.1
urllib3==1.26.13
python-dotenv==0.21.0
python-telegram-bot==13.8